requirements.txt
aiodns==3.5.0
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioretry==6.3.1
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
ccxt==4.4.92
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.3
propcache==0.3.2
pycares==4.9.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1

environs~=14.2.0
betterlogging~=1.0.0
SQLAlchemy~=2.0.41
typing_extensions~=4.14.1